apps/api/requirements/dev.txt
-r base.txt

# Dev / test only
ruff==0.4.4
mypy==1.10.0
django-stubs[compatible-mypy]==5.0.0
djangorestframework-stubs[compatible-mypy]==3.15.0
pytest==8.2.0
pytest-django==4.8.0
pytest-cov==5.0.0
bandit==1.7.8
pip-audit==2.7.3
safety==3.2.0
semgrep==1.71.0
ipython==8.24.0
